Modularize libc bindings into a subproject
