+++ clib/src/main/scala/scala/scalanative/libc/string.scala
@@ -1,5 +1,7 @@
 package scala.scalanative
-package native
+package libc
+
+import native._
 
 @extern object string {
   def strncmp(s1: CString, s2: CString, n: CSize): CInt = extern
