+++ demo/native/smallpt.scala
@@ -1,6 +1,7 @@
 package demo
 
-import scalanative.native._, stdlib._
+import scalanative.native._
+import scalanative.libc.stdlib._
 import java.lang.Math.{PI, sin, cos, abs, pow, sqrt}
 
 @struct
